Unpause connector on subscription
